refactor third-parties menu
