+++ @xen-orchestra/web-core/lib/components/menu/MenuList.vue
@@ -93,7 +93,7 @@ const open = (event: MouseEvent) => {
   border-radius: 0.4rem;
   background-color: var(--color-neutral-background-primary);
   z-index: 1010;
-  overflow: scroll;
+  overflow: auto;
 
   &.horizontal {
     flex-direction: row;
