+++ @xen-orchestra/web/src/components/third-parties/ThirdParties.vue
@@ -3,36 +3,24 @@
     <template #trigger="{ open }">
       <UiDropdownButton @click="open($event)">{{ t('third-parties') }}</UiDropdownButton>
     </template>
-    <DropdownTitle>
-      <UiEasyvirtLogo class="logo" />{{ t('provider-solutions', { provider: 'EasyVirt' }) }}
-    </DropdownTitle>
+    <VtsDropdownTitle>
+      <EasyvirtLogo class="logo" />{{ t('provider-solutions', { provider: 'EasyVirt' }) }}
+    </VtsDropdownTitle>
 
     <MenuItem>
-      <UiLink
-        v-if="dcScopeIp !== undefined"
-        size="medium"
-        :href="`https://${dcScopeIp}`"
-        rel="noopener noreferrer"
-        target="_blank"
-      >
+      <UiLink v-if="dcScopeIp !== undefined" class="link" size="medium" :href="`https://${dcScopeIp}`">
         {{ t('open-app', { name: 'DC Scope' }) }}
       </UiLink>
-      <UiLink v-else size="medium" href="/#/hub/recipes" rel="noopener noreferrer" target="_blank">
+      <UiLink v-else class="link" size="medium" :href="hubLink">
         {{ t('install-app', { name: 'DC Scope' }) }}
       </UiLink>
     </MenuItem>
 
     <MenuItem>
-      <UiLink
-        v-if="dcNetscopeIp !== undefined"
-        size="medium"
-        :href="`https://${dcNetscopeIp}`"
-        rel="noopener noreferrer"
-        target="_blank"
-      >
+      <UiLink v-if="dcNetscopeIp !== undefined" class="link" size="medium" :href="`https://${dcNetscopeIp}`">
         {{ t('open-app', { name: 'DC Netscope' }) }}
       </UiLink>
-      <UiLink v-else size="medium" href="/#/hub/recipes" rel="noopener noreferrer" target="_blank">
+      <UiLink v-else class="link" size="medium" :href="hubLink">
         {{ t('install-app', { name: 'DC Netscope' }) }}
       </UiLink>
     </MenuItem>
@@ -40,19 +28,25 @@
 </template>
 
 <script lang="ts" setup>
+import EasyvirtLogo from '@/components/third-parties/easyvirt/EasyvirtLogo.vue'
+import { useXoRoutes } from '@/remote-resources/use-xo-routes'
 import { useXoVmCollection } from '@/remote-resources/use-xo-vm-collection.ts'
-import { type XoVm } from '@/types/xo/vm.type'
-import DropdownTitle from '@core/components/dropdown/DropdownTitle.vue'
+import VtsDropdownTitle from '@core/components/dropdown/VtsDropdownTitle.vue'
 import MenuItem from '@core/components/menu/MenuItem.vue'
 import MenuList from '@core/components/menu/MenuList.vue'
 import UiDropdownButton from '@core/components/ui/dropdown-button/UiDropdownButton.vue'
-import UiEasyvirtLogo from '@core/components/ui/easyvirt-logo/UiEasyvirtLogo.vue'
 import UiLink from '@core/components/ui/link/UiLink.vue'
+import type { XoVm } from '@vates/types'
 import { computed } from 'vue'
 
 import { useI18n } from 'vue-i18n'
 
 type Product = 'dcscope' | 'dcnetscope'
+
+const { t } = useI18n()
+
+const { vms } = useXoVmCollection()
+const { routes } = useXoRoutes()
 
 function getInstallTime(vm: XoVm, product: Product): string | undefined {
   return vm.other[`xo:${product}:installTime`]
@@ -77,30 +71,23 @@ function getLatestVmIp(vms: XoVm[], product: Product): string | undefined {
   return latestVm?.mainIpAddress
 }
 
-const { t } = useI18n()
-
-const { vms } = useXoVmCollection()
-
 const dcScopeVms = computed(() => vms.value.filter(vm => vm.other['xo:dcscope:installTime'] !== undefined))
 const dcNetscopeVms = computed(() => vms.value.filter(vm => vm.other['xo:dcnetscope:installTime'] !== undefined))
 
 const dcScopeIp = computed(() => getLatestVmIp(dcScopeVms.value, 'dcscope'))
 const dcNetscopeIp = computed(() => getLatestVmIp(dcNetscopeVms.value, 'dcnetscope'))
+
+const hubLink = computed(() => `${routes.value?.xo5}#/hub/recipes`)
 </script>
 
 <style lang="postcss" scoped>
-.ui-link {
+.link {
   text-decoration: none;
-  padding-block: 1.15rem;
+  color: var(--color-brand-txt-base);
+  padding-block: 1.2rem;
   flex-grow: 1;
 }
-.dropdown-title {
-  text-transform: uppercase;
-  font-weight: 600;
-  letter-spacing: 4%;
-  height: 3.6rem;
-  margin: -0.4rem -0.4rem 0 -0.4rem;
-}
+
 .logo {
   vertical-align: middle;
   margin-right: 0.8rem;
