+++ @xen-orchestra/web/src/components/third-parties/easyvirt/EasyvirtLogo.vue
